### 2020-02-22 #2113 @whedon
[PRE REVIEW]: BiRank: Fast and Flexible Ranking on Bipartite Networks with R and Python
**Submitting author:** @yangkcatiu (<a href="http://orcid.org/0000-0003-4627-9273">Kai-Cheng Yang</a>)
**Repository:** <a href="https://github.com/BrianAronson/birankr" target ="_blank">https://github.com/BrianAronson/birankr</a>
**Version:** v0.1.1
