+++ paper.bib
@@ -102,17 +102,6 @@
   year={2002}
 }
 
-@article{moody2004structure,
-  title={The structure of a social science collaboration network: Disciplinary cohesion from 1963 to 1999},
-  author={Moody, James},
-  journal={American sociological review},
-  volume={69},
-  number={2},
-  pages={213--238},
-  year={2004},
-  publisher={Sage Publications Sage CA: Los Angeles, CA}
-}
-
 @article{lehmann2008biclique,
   title={Biclique communities},
   author={Lehmann, Sune and Schwartz, Martin and Hansen, Lars Kai},
