### 2020-02-24 #2113 @drvinceknight
@ati-ozgur @camillescott @Nick-gauthier would you be able to review this software paper for JOSS? It's an interesting piece of graph theory work that is implemented in two languages both R and Python.

If you have not reviewed for JOSS before, you can find information about the process here https://openjournals.readthedocs.io/en/latest/reviewer_guidelines.html and I'd be happy to answer any questions you may have.
