### 2020-03-23 32c0e8f
Updated links in Readme to use absolute paths for links to abide by CRAN standards.
