### 2020-03-23 9577a06
Updated cran-comments to address issues for release of birankr 1.01.
