+++ cran-comments.md
@@ -1,17 +1,26 @@
 This resubmission addresses the following concerns:
 
-> If there are references describing the methods in your package, please
-> add these in the Description field of your DESCRIPTION file in the form
-> authors (year) doi:...
-> authors (year) arXiv:...
-> authors (year, ISBN:...)
-> with no space after 'doi:', 'arXiv:' and angle brackets for auto-linking.
+>    Found the following (possibly) invalid URLs:
+>      URL: https://cran.r-project.org/web/packages/birankr/index.html
+>        From: README.md
+>        Status: 200
+>        Message: OK
+>        CRAN URL not in canonical form
+>      The canonical URL of the CRAN page for a package is
+>        https://CRAN.R-project.org/package=pkgname
+> 
+>    Found the following (possibly) invalid file URIs:
+>      URI: examples/Marvel_social_network.md
+>        From: README.md
+>      URI: birankr.pdf
+>        From: README.md
+>      URI: examples/Marvel_social_network.ipynb
+>        From: README.md
+>      URI: birankpy/README.md
+>        From: README.md
 
-Done. I have added added one reference to the DESCRIPTION file, and I have reformatted function-specific references to match the format outlined above. 
+Done. I have changed links in the readme to use absolute path names and I provided the canonical URL for the CRAN page. 
 
-> Also the \value tags have not been added.
-
-Done. I have added \value tags with descriptions of each functions' output. 
 
 ## Test environments
 * local Windows 10 install, R 3.4.1
