### 2020-04-25 #2113 @arfon
:wave: @rodogi @szhorvat @gvegayon - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
