### 2020-04-25 #2113 @szhorvat
@arfon I'm still behind with a review for JOSS, which turned out to be more work than expected, so I cannot formally commit this time. It looks interesting though, so if time permits, I might post a few comments in the review thread.
