### 2020-04-25 #2113 @szhorvat
> and apologies for not checking first to see if you were already actively reviewing for us

I think it's okay to ask. We can always refuse, and as I said, it does look interesting.
