### 2020-05-04 #2113 @drvinceknight
@camillescott would you be able to review this paper for JOSS? Your indicated research interests include graph theory so I think you will hopefully find it interesting.
