### 2020-05-04 #2113 @drvinceknight
@ati-ozgur would you be able to review this paper for JOSS? You also indicated algorithms and network science as your interests so I hope you'd find the paper of interest :)
