### 2020-06-15 BrianAronson/birankr#19 @Nikoleta-v3
set CI
Setup a CI for the test suite.

There are several CIs that work well with GitHub. Recently I have been using [GitHub actions](https://help.github.com/en/actions/configuring-and-managing-workflows/configuring-a-workflow). Another commonly used CI is [Travis](https://docs.travis-ci.com/user/tutorial/#to-get-started-with-travis-ci-using-github).
