### 2020-06-15 BrianAronson/birankr#20 @Nikoleta-v3
instructions for running tests
Include instructions on the README on how to run the tests implemented for the package.
