### 2020-06-15 BrianAronson/birankr#21 @Nikoleta-v3
`install_requires` requirements bounds
It's best practice to indicate any known lower or upper bounds to the requirements listed in the `install_requires` (in the `setup.py` file) .

For example:

```
install_requires=[
    'numpy>=1.17.0'
    ...
]
```

Adding lower bounds to the requirements is encouraged.
