### 2020-06-15 #2315 @Nikoleta-v3
Hello @yangkcatiu 👋 I am one of the reviewers. I am currently completing the **Documentation** checklist. I opened the following issues:

- https://github.com/BrianAronson/birankr/issues/19
- https://github.com/BrianAronson/birankr/issues/20
- https://github.com/BrianAronson/birankr/issues/21

Please comment on the issues if you have any questions.
