### 2020-06-15 BrianAronson/birankr#20 @yangkcatiu
@BrianAronson Please add instructions for the R part, I'll take care of the python part.
