### 2020-06-15 212e8b2
Add instruction for runing python unit tests
