### 2020-06-15 BrianAronson/birankr#20 @yangkcatiu
https://github.com/BrianAronson/birankr/commit/212e8b2dd3ff9a13892a1fd6b44b62c36ede0c51 adds the instruction for running unit tests for the python version.
