### 2020-06-15 5c2d7c9
Update README.md
Provide instructions for running unit tests
