### 2020-06-15 80b68b9
add version requirements for the dependencies
