+++ setup.py
@@ -5,7 +5,7 @@ with open("pypi_readme.md", "r") as fh:
     long_description = fh.read()
 
 setup(name='birankpy',
-      version='0.1.1',
+      version='0.1.2',
       description = 'Ranking nodes in bipartite networks with efficiency and flexibility',
       long_description=long_description,
       long_description_content_type='text/markdown',
@@ -14,10 +14,10 @@ setup(name='birankpy',
       url="https://github.com/BrianAronson/birankr",
       license = 'MIT',
       install_requires=[
-          'networkx',
-          'pandas',
-          'numpy',
-          'scipy'
+          'networkx==2.2',
+          'pandas>=0.23.4',
+          'numpy>=1.16.2',
+          'scipy>=1.2.0'
       ],
       packages = ['birankpy']
 )
