### 2020-06-15 BrianAronson/birankr#20 @BrianAronson
https://github.com/BrianAronson/birankr/commit/5c2d7c9e1ff5a47d60609fc6d36d2a067090c145 adds instructions for running unit tests for the R version.
