### 2020-06-15 #2315 @gvegayon
Sure! Happy to help. I won't start working on this until the end of this
month since I'm defending my Ph.D. soon (sorry!). I'll start with the
review ASAP.

George
<https://ggvy.cl>

On Mon, Jun 15, 2020 at 7:53 AM Kaicheng(Kevin) Yang <
notifications@github.com> wrote:

> 👋 @gvegayon <https://github.com/gvegayon> thank you for agreeing to
> review for our work, we really appreciate it and low forward to your
> feedback!
>
