### 2020-06-15 BrianAronson/birankr#21 @yangkcatiu
https://github.com/BrianAronson/birankr/commit/80b68b95c1e23dd9695dbb534664a7fbdb653947 should address this issue.
