### 2020-06-15 d21bc33
Specify working directory for the tests
