+++ .github/workflows/python-package.yml
@@ -28,5 +28,6 @@ jobs:
         python -m pip install --upgrade pip
         pip install -e ./
     - name: Run unit tests
+      working-directory: tests
       run: |
-        python tests/test_birankpy.py
+        python test_birankpy.py
