### 2020-06-15 BrianAronson/birankr#19 @yangkcatiu
Github actions for the python version is ready.
