### 2020-06-15 ce731d6
Update README.md
Add travis CI badge for R version
