### 2020-06-15 0d8ab57
Update README.md
Added badge indicating R downloads
