### 2020-06-15 5d84c21
remove r-package.ymal
