+++ .github/workflows/r-package.yml
@@ -1,36 +1,0 @@
-# This is a basic workflow to help you get started with Actions
-
-name: R package
-
-# Controls when the action will run. Triggers the workflow on push or pull request
-# events but only for the master branch
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-
-# A workflow run is made up of one or more jobs that can run sequentially or in parallel
-jobs:
-  # This workflow contains a single job called "build"
-  build:
-    # The type of runner that the job will run on
-    runs-on: ubuntu-latest
-
-    # Steps represent a sequence of tasks that will be executed as part of the job
-    steps:
-    # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-    - uses: actions/checkout@v2
-    
-    # Set up R environment
-    - uses: r-lib/actions/setup-r@master
-      with:
-          r-version: '3.6.1'
-          
-    # Install devtools
-    - name: 
-      run: Rscript -e 'install.packages("devtools")'
-
-    # Runs a single command using the runners shell
-    - name: Run tests
-      run: Rscript -e "devtools::test()"
