### 2020-06-15 BrianAronson/birankr#19 @BrianAronson
Travis CI for the R version is ready.
