### 2020-06-15 BrianAronson/birankr#20 @Nikoleta-v3
Thank you both for addressing my comments.
