### 2020-06-17 BrianAronson/birankr#23 @Nikoleta-v3
JOSS:  state of the field & statement of need 
> State of the field: Do the authors describe how this software compares to other commonly-used packages?

You very nicely compare your methodolody to the common one-mode projections approach. However, the paper is missing a few sentences explaining how the software compares to other packages.

This bring me to:

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

You clearly state the problems the software is designed to solve, but not the target audience. Following my comment above I think you should also add a sentence that you specify this.
