### 2020-06-17 BrianAronson/birankr#24 @Nikoleta-v3
JOSS: minor edit comments
>  we present BiRank, a R and Python package that performs PageRank

`an R` package? I am not 100% sure.

> but in many applications, we tend to prefer CoHITS over HITS as CoHITS ranks are less influenced by the presence of outliers with extreme degrees.

Is this a personal preference or a field preference? If it's a field preference please reference.
