### 2020-06-17 #2315 @Nikoleta-v3
👋 @yangkcatiu.  I am now completing the **Software paper** checklist. I opened the following issues:

- https://github.com/BrianAronson/birankr/issues/23
- https://github.com/BrianAronson/birankr/issues/24

Please comment on the issues if you have any questions.
