### 2020-06-17 BrianAronson/birankr#24 @yangkcatiu
Thanks for the feedback. Regarding the first comment, we implemented the same functionality both in R and Python.

The R version is available through: https://cran.r-project.org/web/packages/birankr/index.html
And the Python version available through: https://pypi.org/project/birankpy/

Hope this address your question.
