### 2020-06-17 BrianAronson/birankr#24 @Nikoleta-v3
Sorry my comment was not clear.

In the `paper.md` you have written `we present BiRank, a R package`. I believe it should say `an R package`, you are missing a `n`.
