### 2020-06-17 BrianAronson/birankr#24 @yangkcatiu
Sorry I misunderstood. This is a good catch. I did some googling and believe you are correct, should be `an R package`. https://github.com/BrianAronson/birankr/commit/720c08e87551d4db92c351e8ab99c4c6694dee24 should fix this.
