### 2020-06-17 BrianAronson/birankr#24 @Nikoleta-v3
No worries at all. I should have specified that both my comments are regarding the paper: https://github.com/BrianAronson/birankr/blob/master/paper.md
