### 2020-06-17 BrianAronson/birankr#23 @yangkcatiu
https://github.com/BrianAronson/birankr/commit/76fd35dc2456753ddc5432a2b484cdd751e82460 includes new sentences addressing these issues.
