### 2020-06-18 BrianAronson/birankr#24 @Nikoleta-v3
Ιn the `paper.md` you state that:

> but in many applications, we tend to prefer CoHITS over HITS as CoHITS ranks are less influenced by the presence of outliers with extreme degrees.

Is this a personal preference or a field preference? If it's a field preference please add a reference.
