### 2020-06-21 c5443c0
Update paper.bib
Add socarxiv paper to references
