+++ paper.bib
@@ -126,3 +126,12 @@
   publisher={Nature Publishing Group}
 }
 
+@misc{aronson_yang_odabas_ahn_perry_2020,
+ title={Comparing Measures of Centrality in Bipartite Social Networks: A Study of Drug Seeking for Opioid Analgesics},
+ url={osf.io/preprints/socarxiv/hazvs},
+ DOI={10.31235/osf.io/hazvs},
+ publisher={SocArXiv},
+ author={Aronson, Brian and Yang, Kai-Cheng and Odabas, Meltem and Ahn, Yong Y and Perry, Brea L},
+ year={2020},
+ month={Jun}
+}
