### 2020-06-21 f04d087
Update paper.md
Added reference for comment about our preference of CoHITS over HITS
