### 2020-06-21 BrianAronson/birankr#24 @BrianAronson
Thank you for catching the missing reference. I have updated the manuscript to reference a paper that explicitly recommends CoHITS over HITS. https://github.com/BrianAronson/birankr/commit/f04d0879e0e4552b408e91c5343cf05bacee4807
