### 2020-06-21 BrianAronson/birankr#24 @Nikoleta-v3
Thank you both for addressing my comments.
