### 2020-07-04 BrianAronson/birankr#25 @gvegayon
[JOSS] vertex ids range
This is regarding the JOSS review https://github.com/openjournals/joss-reviews/issues/2315

When reading networks, the function `br_cohits` factors the vertex ids, in particular. In the [example included in the readme file](https://github.com/BrianAronson/birankr/tree/f04d0879e0e4552b408e91c5343cf05bacee4807#example), while the edgelist has elements ranging from 1 through 10,000, the ranking function returns only 6,348 elements, meaning that it took the unique elements instead of taking the whole range. This makes me think regarding what happens when the network includes isolates. I understand that isolates may not have any role here, yet it is important to document this behavior. Another alternative would be adding an argument specifying the range of the vertex ids. I haven't checked the other functions, but I'm guessing it is the same.
