### 2020-07-04 BrianAronson/birankr#26 @gvegayon
[JOSS] test results
This is regarding the JOSS review https://github.com/openjournals/joss-reviews/issues/2315

I see that the R package has tests, but most of them (if not all) are focused on catching errors and equivalences instead of looking at the actual numbers. It would be nice to see that you guys are including tests in which you know what are the exact results, e.g. a real-world (or simulated) network compute the ranks "by hand" (i.e. not using your functions), and compare those results to what your functions return. That would be (a) reassuring for the users, and (b) very nice.
