### 2020-07-04 #2315 @gvegayon
Hi there! I'm back :). I started the review, but I was trying to mark the checklist and it seems that I'm no longer able to do so. I tried re-opening the collab link sent by JOSS but it expedited. What can I do? I'm still reviewing and will let you know when all is done.
