### 2020-07-05 BrianAronson/birankr#25 @yangkcatiu
@BrianAronson would you please take a look at this?
