### 2020-07-05 BrianAronson/birankr#26 @yangkcatiu
@BrianAronson I'll work on the test cases.
