### 2020-07-07 a778191
add new birank test cases
