### 2020-07-07 BrianAronson/birankr#26 @yangkcatiu
@gvegayon In https://github.com/BrianAronson/birankr/pull/27 I created several small toy networks and calculated the ranks by hand for testing the functions. Let me know what you think!
