### 2020-07-07 BrianAronson/birankr#26 @BrianAronson
@yangkcatiu Thanks for this. I will duplicate these in the R package.
