### 2020-07-08 e642880
Add files via upload
Make tolerance check both modes of the network; make output show tolerance checks on both modes of the network
