### 2020-07-08 4fd3b1a
Add files via upload
Revert change to testthat (previous commit based on old file)
