### 2020-07-08 11ec409
Add files via upload
Add 32 unit tests that estimate whether bipartite_rank matches ranks that were computed "by hand" (externally).
