### 2020-07-08 8901d16
Add files via upload
Updated documentation to indicate how functions interact with network isolates and to clarify how functions interact with edge lists.
