### 2020-07-08 BrianAronson/birankr#25 @BrianAronson
@gvegayon, thank you for you feedback! Great suggestion about documenting how the package functions handle isolates - we had forgotten to do so. In addition, I have clarified how the function handles edge lists. I have added this documentation in https://github.com/BrianAronson/birankr/commit/8901d1649fdbfe3488d95119fd5092c99ad38ffb. Below is a detailed explanation for your edification:

## Edge Lists with Vertex Indices

### Intro

When birankr functions are applied to edge lists, birankr always treats vertices as unique names, rather than as indices of vertex positions. We made this decision based on a strong assumption that individuals who input edge lists rather than matrices into birankr's functions are likely to use edge lists with named vertices. For example, below is an imagined typical use case for supplying the function with edge lists:

#### Create data
```
library(birankr)
library(randomNames)
set.seed(1)
df <- data.frame(
  patient_id = sample(x = 1:10000, size = 10000, replace = T),
  provider_id = sample(x = 1:5000, size = 10000, replace = T)
)
head(df)
  patient_id provider_id
1       1017          10
2       8004        3356
3       4775         796
4       9725        2493
5       8462        3556
6       4050        2788
```

#### Replace numeric names with standard names
```
pat_names <- randomNames(n = 10000)
pat_names <- unique(pat_names)[1:length(unique(df$patient_id))]
prov_names <- randomNames(n = 5000)
prov_names <- unique(prov_names)[1:length(unique(df$provider_id))]
pat_index <- data.frame(patient_id = unique(df$patient_id), patient_name = pat_names)
prov_index <- data.frame(provider_id = unique(df$provider_id), provider_name = prov_names)
df <- merge(df, pat_index, by = "patient_id", sort = F)
df <- merge(df, prov_index, by = "provider_id", sort = F)
df <- df[, c("patient_name", "provider_name")]
head(df)
      patient_name           provider_name
1 Bigback, Mikayla Kills In Sight, Traivon
2 Otterbein, Brady Kills In Sight, Traivon
3   Gonzales, Raul Kills In Sight, Traivon
4    Kim, Jaspreet        Bellino, Melissa
5    Kim, Jaspreet        al-Azimi, Abbaad
6     Denny, Carlo        al-Azimi, Abbaad
```

#### Return ranks
```
head(br_cohits(data = df))
      patient_name         rank
1 Bigback, Mikayla 0.0001039978
2 Otterbein, Brady 0.0003383414
3   Gonzales, Raul 0.0001039978
4    Kim, Jaspreet 0.0004623346
5     Denny, Carlo 0.0003335763
6 Macdonald, Wyatt 0.0002307766
```

### Reviewer Recommendation
Some users may prefer that they receive centrality estimates for isolated nodes, and for their output to be ordered according to the edge list indices. A suggestion was to provide an option to each function that allows for edge lists containing vertex indices to be read as vertex indices.

### Problem
Edge lists reflecting vertex indices do not always indicate the full dimensions of a network matrix. For example, if we made the functions assume that the input network's dimensions are equivalent to the maximum vertex found in each mode of the edge list, this might lead to incorrect estimates of the network's dimensions (e.g. if maximum edge list index in top mode is 9999 but number of nodes in top mode = 10000 and column 10000 is an isolate). If I create an option allowing the package to treat edge lists as vertex indices rather than named vertex IDs, I would also need to add an option allowing users to specify the dimensions of the underlying network. However, I think users would have a much easier time transforming their edge list into a matrix before supplying it to the network than inputting their edge list into our function and adding two new arguments.

### Solution

We now have the package documentation describe how each function interacts with edge lists more explicitly, and we provide documentation that recommends users to transform their data into a matrix when they desire their output to include vertex isolates. For example, users could do the following for the above use-case:

#### Create data
```
library(birankr)
set.seed(1)
df <- data.frame(
  patient_id = sample(x = 1:10000, size = 10000, replace = T),
  provider_id = sample(x = 1:5000, size = 10000, replace = T)
)
```
#### Convert to sparse matrix
```
df <- sparseMatrix(i = df$patient_id, j = df$provider_id, x = 1)
dim(df)
[1] 10000  5000
head(br_cohits(data = df))
[1] 1.500000e-05 1.500000e-05 6.601004e-05 7.668377e-05 1.500000e-05
[6] 1.500000e-05
```

## Clarify How Functions Interact with Isolates
The bipartite ranking functions work with networks that contain isolates. Specifically, birankr assigns isolates a value of  (1 - alpha) / (n_columns) or (1 - beta) / (n_nrows) depending on their mode in the network. These values will always be smaller than any value assigned to non-isolates. Returning a number rather than NAs was a judgment call that we made. We figured that it is likely to be more useful for users to have a very small non-zero value instead. However, we now note in the package documentation that caution should be used for such cases. Below is an example:

#### Create data
```
library(birankr)
set.seed(1)
df <- data.frame(
  patient_id = sample(x = 1:10000, size = 10000, replace = T),
  provider_id = sample(x = 1:5000, size = 10000, replace = T)
)
```
#### Convert to sparse matrix
```
df <- sparseMatrix(i = df$patient_id, j = df$provider_id, x = 1)
```

#### Check that the above sparse matrix includes isolates.
```
table(rowSums(df))
0    1    2    3    4    5    6    8
3684 3681 1833  601  166   26    8    1
table(colSums(df))
0    1    2    3    4    5    6    7    8    9
657 1373 1380  878  451  171   66   16    6    2
```

#### Illustrate how birankr assigns values to isolates.
```
(1 - .85) * (1 / nrow(df))
[1] 1.5e-05
cohits <- br_cohits(df)
min(cohits)
[1] 1.5e-05
data.frame(id = 1:6, degree = head(rowSums(df)), cohits = head(cohits))
  id degree       cohits
1  1      0 1.500000e-05
2  2      0 1.500000e-05
3  3      1 6.601004e-05
4  4      1 7.668377e-05
5  5      0 1.500000e-05
6  6      0 1.500000e-05
```
