### 2020-07-08 BrianAronson/birankr#26 @BrianAronson
@gvegayon thank you for your suggestion. The following commit adds tests per your recommendation in the R version of the package. https://github.com/BrianAronson/birankr/commit/11ec4096e2a8afe0c8cd63ca8d55709d2858172f
