### 2020-07-09 BrianAronson/birankr#28 @gvegayon
[JOSS] R CMD check notes
This is regarding the JOSS review https://github.com/openjournals/joss-reviews/issues/2315

Running `R CMD check --as-cran` after compiling the package returns the following issues:

```
Found the following (possibly) invalid URLs:
  URL: https://doi.org/10.1145/1291233.1291378
    From: man/br_bgrm.Rd
    Status: 403
    Message: Forbidden
  URL: https://doi.org/10.1145/1557019.1557051
    From: man/br_cohits.Rd
    Status: 403
    Message: Forbidden
```

This is an odd one since hidden files should be excluded automatically. Perhaps think about adding it to .Rbuildignore?

```
Found the following hidden files and directories:
  .github
```
