### 2020-07-09 #2315 @gvegayon
Hey! So everything looks good on my side. I tested the package, compiled it, and run the checks using R CMD check. The authors addressed a couple of key issues that submitted regarding testing and documentation. Issue 28 is regarding some R CMD check notes that are very easy to address, so with that, we are good to go!
