### 2020-07-09 BrianAronson/birankr#28 @BrianAronson
Nice catch and strange error... Thanks! I will look into this.
