### 2020-07-09 9bfed0c
Update .Rbuildignore
Ignore .Rd files that are throwing unnecessary notes from R CMD check --as-cran
