### 2020-07-09 BrianAronson/birankr#28 @BrianAronson
@gvegayon Thanks for your suggested solution. I have implemented it in https://github.com/BrianAronson/birankr/commit/9bfed0ca759a200eb1b5a55e8796a5e4bbabf4a7.
