### 2020-07-10 9a27455
Make minor type edits.
