### 2020-07-10 #2315 @drvinceknight
> @drvinceknight we have addressed the issue.

Thanks, I've made some minor type edits in https://github.com/BrianAronson/birankr/pull/29

Once that's merged could you:

Make a tagged release and archive, and report the version number and archive DOI here. Please ensure that the archive deposit has the correct metadata  so that the title and author list match the paper.
