### 2020-07-10 #2315 @yangkcatiu
We appreciate the edits!

- [x] Merge the edits
- [x] Make a tagged release: v1.0 release at https://github.com/BrianAronson/birankr/releases
- [x] Make an archive: DOI 10.5281/zenodo.3939045
